Create D.3-mix-audio.maxpat
Here's the audio mix patch with a little bit of spatialization. I panned all the channels according to the 3x3 grid. Centre video always plays in the centre and the others are organized in a circle with surrounding the centre one. There's a parameter [0. - 1.] sent to "s rota" which rotates the whole spatialization setting.